backend/requirements.txt
flask==2.0.1
Flask==2.0.1
requests==2.26.0
Werkzeug==2.0.3

# Base----------------------------------
albumentations>=1.1.0
bio==1.6.2
matplotlib
opencv-python>=4.1.1.26
opencv-python-headless>=4.1.1.26
Pillow
PyYAML
scikit-image
scikit-learn
scipy
torch>=1.10.0
torchvision>=0.11.0
numpy<2
protobuf<=3.20.1
pandas
mysql-connector-python
python-dotenv
sbol3
sbol_utilities
schedule


# Logging----------------------------------
wandb
tensorboard

# Extras----------------------------------
pycocotools>=2.0.2
setuptools==59.5.0
flask-cors
